refactor(contracts): move page number formatting
